Disable baseline snapping, and enable rounding to integers for scroll offsets again
The baseline snapping needs to be disabled so that the lines are
properly placed, and not snapped to pixels. While using integer based
scroll offsets greatly reduces the pressure on the Skia font cache.

+++ src/renderer/mod.rs
@@ -412,7 +412,7 @@ impl Renderer {
                 result.font_changed = true;
             }
             DrawCommand::LineSpaceChanged(new_linespace) => {
-                self.grid_renderer.update_linespace(new_linespace as f32);
+                self.grid_renderer.update_linespace(new_linespace);
                 result.font_changed = true;
             }
             DrawCommand::DefaultStyleChanged(new_style) => {
